rename: 구조 변경
