--- app/(auth)/login/_modules/index.ts
@@ -1,1 +1,0 @@
-export { LoginPageView } from "./LoginPageView";
