+++ entities/user/apis/apis.ts
@@ -1,0 +1,10 @@
+import { fetchClient } from "@/shared/fetch";
+import type { UserProfileResponse } from "@/entities/user/models/types";
+
+export const getUserProfile = async (): Promise<UserProfileResponse> => {
+  const response = await fetchClient.get<UserProfileResponse>(
+    "user-profiles/me"
+  );
+
+  return response.data.data;
+};
