+++ modules/auth/ui/index.ts
@@ -1,0 +1,1 @@
+export { LoginPageView } from "./login-page-view";
